% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snapverse.R
\docType{package}
\name{snapverse}
\alias{snapverse}
\alias{snapverse-package}
\title{snapverse: The primary sector package for the SNAPverse.}
\description{
The snapverse package loads a subset of individual SNAPverse member packages.
}
\details{
\describe{
  \item{snapstat}{functions for working with popular, curated SNAP data from SNAPverse data packages.}
  \item{snapplot}{functions for graphing SNAP data.}
  \item{snaplocs}{functions for working with SNAP point locations data.}
  \item{snapprep}{functions for SNAP source data extraction, manipulation and curation for various projects.}
  \item{alfresco}{functions for working with and analyzing ALFRESCO wildfire model inputs and outputs.}
  \item{snapclim}{SNAP climate data.}
  \item{snapfire}{SNAP fire data.}
  \item{snappoly}{collection of popular polygon shapefile data sets used at SNAP.}
  \item{snapgrid}{collection of popular gridded maps used at SNAP.}
  \item{snapdist}{SNAP spatial distributions examples data set.}
  \item{snapapps}{curated collection of SNAP R Shiny apps loadable via R package.}
  \item{snapdash}{selected SNAP Shiny Dashboard apps.}
  \item{snapflex}{selected SNAP Flex Dashboard apps.}
  \item{snapdocs}{other SNAP interactive documents.}
}
}
