% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snapverse.R
\name{snapverse_update}
\alias{snapverse_update}
\title{Update \code{snapverse}}
\usage{
snapverse_update(force = FALSE, quiet = TRUE)
}
\arguments{
\item{force}{If \code{TRUE}, force member install even if unchanged.}

\item{quiet}{logical.}
}
\description{
Update \code{snapverse} and its member packages.
}
\details{
This function updates packages from GitHub. See \url{https://leonawicz.github.io/sv}.
}
\examples{
\dontrun{
snapverse_update()
}
}
